requirements.txt
telethon==1.34.0
python-telegram-bot==20.7
python-dotenv==1.0.0
openai==1.12.0
schedule==1.2.0
requests==2.31.0
python-dateutil==2.8.2
